+++ app/gui/view/graph-editor/src/lib.rs
@@ -2600,7 +2600,7 @@ fn new_graph_editor(app: &Application) -> GraphEditor {
     }
 
 
-    // === Edited node growth/shrink animation. ===
+    // === Edited node growth/shrink animation ===
 
     component::node::growth_animation::initialize_edited_node_animator(&model, &frp, scene);